+++ app/src/main/java/de/kaffeemitkoffein/tinyweatherforecastgermany/WeatherIcons.java
@@ -149,8 +149,7 @@ public class WeatherIcons {
             case SYMBOL_SUN: result = R.mipmap.symbol_sun; break;
             case PIN : result = R.mipmap.pin; break;
             case IC_LAUNCHER_BW : result = R.mipmap.ic_launcher_bw; break;
-            case GERMANY: result = R.drawable.germany; break;
-            case MAP_COLLAPSED: result = R.drawable.map_collapsed; break;
+            case GERMANY: result = R.drawable.germany2; break;
             case RADARINFOBAR: result = R.drawable.radarinfobar; break;
             case RADIO_BUTTON_UNCHECKED: result = R.mipmap.ic_radio_button_unchecked_white_24dp; break;
             case RADIO_BUTTON_CHECKED: result = R.mipmap.ic_radio_button_checked_white_24dp; break;
@@ -197,8 +196,6 @@ public class WeatherIcons {
                 case WIND_BEAUFORT_11 : result = R.mipmap.wind_beaufort_11_black; break;
                 case WIND_BEAUFORT_12 : result = R.mipmap.wind_beaufort_12_black; break;
                 case SYMBOL_SUN: result = R.mipmap.symbol_sun_black; break;
-                case GERMANY: result = R.drawable.germany_black; break;
-                case MAP_COLLAPSED: result = R.drawable.map_collapsed_black; break;
                 case RADIO_BUTTON_UNCHECKED: result = R.mipmap.ic_radio_button_unchecked_black_24dp; break;
                 case RADIO_BUTTON_CHECKED: result = R.mipmap.ic_radio_button_checked_black_24dp; break;
                 case IC_INFO_OUTLINE: result = R.mipmap.ic_info_outline_black_24dp; break;