+++ CHANGELOG.md
@@ -6,13 +6,15 @@
 - fix wind speed in overview chart displayed too high when temperature was below zero
 - fixed possible crash related to the overview chart when data about the precipitation amount was missing
 - added optional display of the precipitation amount to the overview chart
-- switch the rain radar layers from WGS84 (EPSG 4326) to mercator projection (EPSG 3857) to fit the target map, thus displaying the rain radar with a better local accuracy
+- switch the rain radar layers from WGS84 (EPSG 4326) to Mercator projection (EPSG 3857) to fit the target map, thus displaying the rain radar with a better local accuracy
 - Gadgetbridge support: fix wrong moon-phase in daily forecast
 - turning the Gadgetbridge support off/on now always triggers an update
 - fix some preferences missing when saving the whole preference instance
 - fix wind unit in overview chart km/h instead of kn
 - moved the wind options for the overview chart from "wind" to "overview chart" in the settings
 - bold widget: align the current day to the left to prevent line breaks in temperature display when the values have two digits
+- switched weather warnings activity map to use a static OpenStreetMap in Mercator Projection
+- rewritten plot functions to draw warning polygons and administrative borders on Mercator projection (EPSG 3857)
 
 ## Version 0.62.0: